+++ src/utils/jm_domain_detect.py
@@ -216,9 +216,8 @@ def jm_auto_spider_img_thread():
         for spider_image_keyword_item in spider_img_keyword_detail:
             logger.debug("Current spider kew word: " + str(spider_image_keyword_item.strip()))
             try:
-                search_download_jm(spider_image_keyword_item.strip())
-                if not constants.JM_SD_auto_flag:
-                    # logger.warning("stop jm spider.")
+                if not search_download_jm(spider_image_keyword_item.strip()):
+                    logger.success("Stop jm spider finished")
                     return False
             except Exception as e:
                 logger.error(f"Unknown error, detail: {e}")